Implemented a Chrome fix, CSS improvements and TVMaze episode linking functionality added
changed the way the event listener worked for the select element as it was working fine in Firefox but didn't work in Chrome.
Made many improvements to the CSS, however as the project is far from finished there is still a lot I intend to do.
Added click functionality to each episode card that opens its page on TVMaze.
Made several improvements to the code structure.